Made the addition button functional The addition button does simple integer addition and displays the result in the resultField.

+++ JavaCalculator/src/Operations.java
@@ -1,17 +1,17 @@
 public class Operations {
-    public double addition(double x, double y) {
+    public double addition(int x, int y) {
         return x + y;
     }
-    public double subtraction(double x, double y) {
+    public double subtraction(int x, int y) {
         return x - y;
     }
-    public double multiplication(double x, double y) {
+    public double multiplication(int x, int y) {
         return x * y;
     }
-    public double division(double x, double y) {
-        return x / y;
+    public double division(int x, int y) {
+        return (double) x / y;
     }
-    public String addNumber(double num, String result) {
+    public String addNumber(int num, String result) {
         result = result + num;
         return result;
     }
